+++ src/Core/Dfe.Complete.Application/Projects/Models/ProjectDto.cs
@@ -86,7 +86,9 @@ namespace Dfe.Complete.Application.Projects.Models
 
         public  ICollection<Contact> Contacts { get; set; } = new List<Contact>();
 
-        public  ICollection<Note> Notes { get; set; } = new List<Note>();
+        public  ICollection<NoteDto> Notes { get; set; } = new List<NoteDto>();
+        
+        public  ICollection<SignificantDateHistoryDto> SignificantDateHistories { get; set; } = new List<SignificantDateHistoryDto>();
 
         public User RegionalDeliveryOfficer { get; set; } = default!;