+++ src/Core/Dfe.Complete.Application/Projects/Models/SignificantDateHistoryReasonDto.cs
@@ -1,0 +1,16 @@
+﻿using Dfe.Complete.Domain.ValueObjects;
+
+namespace Dfe.Complete.Application.Projects.Models;
+
+public class SignificantDateHistoryReasonDto
+{
+    public SignificantDateHistoryReasonId Id { get; set; }
+
+    public DateTime CreatedAt { get; set; }
+
+    public DateTime UpdatedAt { get; set; }
+
+    public string? ReasonType { get; set; }
+
+    public SignificantDateHistoryId? SignificantDateHistoryId { get; set; }
+}