+++ CHANGELOG.md
@@ -22,7 +22,8 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 
-### Added  
+### Added
+- New route `/projects/{projectId}/date-history`
 
 ### Changed